+++ core/src/main/java/org/jdbi/v3/core/transaction/SerializableTransactionRunner.java
@@ -79,19 +79,6 @@ public class SerializableTransactionRunner extends DelegatingTransactionHandler 
     }
 
     @Override
-    public <R, X extends Exception> R inTransaction(Handle handle,
-                                                    TransactionIsolationLevel level,
-                                                    HandleCallback<R, X> callback) throws X {
-        final TransactionIsolationLevel initial = handle.getTransactionIsolationLevel();
-        try {
-            handle.setTransactionIsolationLevel(level);
-            return inTransaction(handle, callback);
-        } finally {
-            handle.setTransactionIsolationLevel(initial);
-        }
-    }
-
-    @Override
     public TransactionHandler specialize(Handle handle) throws SQLException {
         return new SerializableTransactionRunner(getDelegate().specialize(handle));
     }